move non-gatekeeping check
