+++ tsutils/__init__.py
@@ -1,10 +1,10 @@
 from .cog_settings import *
-from .emoji_utils import *
-from .format_utils import *
+from .emoji import *
+from .formatting import *
 from .helper_classes import *
 from .helper_functions import *
 from .json_utils import *
 from .menu import *
-from .pad_utils import *
-from .tsubaki_utils import *
+from .pad import *
+from .tsubaki import *
 from .user_interaction import *
