+++ tsutils/emoji.py
